refactor(wallet): CORS consistency, formatTokenBalance dedup, NFT race condition docs
CORS headers — single source of truth:
- Add apps/web/src/app/api/wallet/_cors.ts with WALLET_CORS_HEADERS
  constant and walletOptionsResponse() helper
- All three wallet routes (tokens, nfts, transactions) now import from
  there — a single edit to _cors.ts propagates to all endpoints

formatTokenBalance — shared util replaces two diverging copies:
- Add formatTokenBalance(rawBalance, decimals, maxDecimals=6) to
  packages/shared/src/utils/wallet.ts
- WalletOverview.tsx: remove local formatBalance, use shared with
  maxDecimals=4 (overview card precision)
- SendModal.tsx: remove local formatBalance, use shared with
  maxDecimals=6 (send modal needs finer precision for small amounts)
- Both callers now have identical logic; the precision difference is
  expressed explicitly at the call site

NFT ownership race condition — design decision documented:
- Add comment above the nftOwnership query in sendNftActionImpl
  explaining: check is a UX guard (not a security gate), true ownership
  enforcement is on-chain via safeTransferFrom revert, the sub-second
  race window is intentionally accepted, and a failed log on on-chain
  revert is correct audit behavior

Made-with: Cursor
