+++ apps/web/src/app/api/wallet/nfts/route.ts
@@ -28,14 +28,10 @@ import type { NextRequest } from 'next/server';
 import { NextResponse } from 'next/server';
 import { isAddress } from 'viem';
 
-const CORS_HEADERS = {
-  'Access-Control-Allow-Origin': '*',
-  'Access-Control-Allow-Methods': 'GET, OPTIONS',
-  'Access-Control-Allow-Headers': 'Content-Type, Authorization',
-};
+import { walletOptionsResponse } from '../_cors';
 
 export function OPTIONS() {
-  return new Response(null, { status: 204, headers: CORS_HEADERS });
+  return walletOptionsResponse();
 }
 
 export const GET = withErrorHandling(async (request: NextRequest) => {
