+++ apps/web/src/app/api/wallet/tokens/route.ts
@@ -36,14 +36,10 @@ import {
   parseAbi,
 } from 'viem';
 
-const CORS_HEADERS = {
-  'Access-Control-Allow-Origin': '*',
-  'Access-Control-Allow-Methods': 'GET, OPTIONS',
-  'Access-Control-Allow-Headers': 'Content-Type, Authorization',
-};
+import { walletOptionsResponse } from '../_cors';
 
 export function OPTIONS() {
-  return new Response(null, { status: 204, headers: CORS_HEADERS });
+  return walletOptionsResponse();
 }
 
 const publicClient = createPublicClient({
