+++ apps/web/src/app/api/wallet/transactions/route.ts
@@ -46,14 +46,10 @@ import { isAddress } from 'viem';
 // unbounded queries for users with large transaction histories.
 const MAX_RECORDS_PER_SOURCE = 500;
 
-const CORS_HEADERS = {
-  'Access-Control-Allow-Origin': '*',
-  'Access-Control-Allow-Methods': 'GET, OPTIONS',
-  'Access-Control-Allow-Headers': 'Content-Type, Authorization',
-};
+import { walletOptionsResponse } from '../_cors';
 
 export function OPTIONS() {
-  return new Response(null, { status: 204, headers: CORS_HEADERS });
+  return walletOptionsResponse();
 }
 
 export const GET = withErrorHandling(async (request: NextRequest) => {
